Create bitmovin sprite lambda
This function creates a sprite for bitmovin encoded videos.  A sprite has two parts, a jpg immage that is really a bunch of thumbnails stiched together sequentially, and a vtt
file that makes video times to locations on the image so it shows the right thumb when you scroll over the corresponding video duration time.  After creating the sprite files, the function finds their url location and uses it to upload them to the video encoding folder on S3.

Tags: bitmovin, sprite, lambda