Parse sprite jpg name from response
When we create the sprite, a jpg and vtt file are generated.  The vtt file references the jpg by the name it was given when created by bitmovin.  We were overwriting the name of the jpg when transfering it to S3, so the vtt no longer pointed to the right place.  This change parses the original name of the jpg from the creation response, and uses it to name the file when sending to S3.  Also took the opportunity to dry up some duplication in the handler.

Tags: bitmovin, sprite, lambda, refactor